requirements.txt
djangorestframework
djangorestframework-simplejwt
Django==4.2.7
psycopg2
